[#52] Split scalastyle in its own subproject
