--- core/src/sbt-test/scalafmt/test/scalastyle-config.xml
@@ -1,83 +1,0 @@
-<scalastyle>
-  <name>kantan configuration</name>
-
-  <!-- = Naming ================================================================================================== -->
-  <!-- =========================================================================================================== -->
-  <check class="org.scalastyle.scalariform.ClassNamesChecker"          level="error"   enabled="true">
-    <parameters>
-      <parameter name="regex">[A-Z\~\*\/][A-Za-z]*</parameter>
-      <parameter name="ignoreRegex">^[^a-zA-Z0-9].*$</parameter>
-    </parameters>
-  </check>
-  <check class="org.scalastyle.scalariform.ObjectNamesChecker"         level="error"   enabled="true">
-    <parameters>
-      <parameter name="regex">[A-Za-z\~\*\/][A-Za-z]*</parameter>
-      <parameter name="ignoreRegex">^[^a-zA-Z0-9].*$</parameter>
-    </parameters>
-  </check>
-  <check class="org.scalastyle.scalariform.ClassTypeParameterChecker"  level="warning" enabled="true">
-    <parameters>
-      <parameter name="regex">^[A-Z_]$</parameter>
-    </parameters>
-  </check>
-  <check class="org.scalastyle.scalariform.MethodArgumentNamesChecker" level="error"   enabled="true">
-    <parameters>
-      <parameter name="regex">^[a-z][A-Za-z0-9]*$</parameter>
-    </parameters>
-  </check>
-  <check class="org.scalastyle.scalariform.MethodNamesChecker"         level="error"   enabled="true">
-    <parameters>
-      <parameter name="regex">^[a-z][a-zA-Z0-9]*$</parameter>
-      <parameter name="ignoreRegex">^([^a-zA-Z0-9].*)|(unary_!)$</parameter>
-      <parameter name="ignoreOverride">true</parameter>
-    </parameters>
-  </check>
-  <check class="org.scalastyle.scalariform.PackageNamesChecker"        level="error"   enabled="true">
-    <parameters>
-      <parameter name="regex">^[a-z][a-z0-9]*$</parameter>
-    </parameters>
-  </check>
-  <check class="org.scalastyle.scalariform.FieldNamesChecker"          level="error"   enabled="true">
-    <parameters>
-      <parameter name="regex">^[a-z][A-Za-z0-9]*$</parameter>
-      <parameter name="objectFieldRegex">^[A-Za-z][A-Za-z0-9]*$</parameter>
-    </parameters>
-  </check>
-
-  <!-- = Misc. =================================================================================================== -->
-  <!-- =========================================================================================================== -->
-  <check class="org.scalastyle.scalariform.CovariantEqualsChecker"           level="error"   enabled="true"/>
-  <check class="org.scalastyle.scalariform.DeprecatedJavaChecker"            level="error"   enabled="true"/>
-  <check class="org.scalastyle.scalariform.EmptyClassChecker"                level="error"   enabled="true"/>
-  <check class="org.scalastyle.scalariform.EmptyInterpolatedStringChecker"   level="error"   enabled="true"/>
-  <check class="org.scalastyle.scalariform.EqualsHashCodeChecker"            level="error"   enabled="true"/>
-  <check class="org.scalastyle.scalariform.ImportOrderChecker"               level="error"   enabled="true">
-    <parameters>
-      <parameter name="groups">all</parameter>
-      <parameter name="group.all">.+</parameter>
-      <parameter name="lexicographic">true</parameter>
-    </parameters>
-  </check>
-  <check class="org.scalastyle.scalariform.NoCloneChecker"                   level="error"   enabled="true"/>
-  <check class="org.scalastyle.scalariform.NoFinalizeChecker"                level="error"   enabled="true"/>
-  <check class="org.scalastyle.scalariform.NotImplementedErrorUsage"         level="error"   enabled="true"/>
-  <check class="org.scalastyle.scalariform.NullChecker"                      level="warning" enabled="true"/>
-  <check class="org.scalastyle.scalariform.OverrideJavaChecker"              level="error"   enabled="true"/>
-  <check class="org.scalastyle.scalariform.ProcedureDeclarationChecker"      level="error"   enabled="true"/>
-  <check class="org.scalastyle.scalariform.PublicMethodsHaveTypeChecker"     level="error"   enabled="true">
-    <parameters>
-      <parameter name="ignoreOverride">true</parameter>
-    </parameters>
-  </check>
-  <check class="org.scalastyle.scalariform.RedundantIfChecker"               level="warning" enabled="true"/>
-  <check class="org.scalastyle.scalariform.ReturnChecker"                    level="error"   enabled="true"/>
-  <check class="org.scalastyle.scalariform.SimplifyBooleanExpressionChecker" level="error"   enabled="true"/>
-  <check class="org.scalastyle.scalariform.StructuralTypeChecker"            level="warning" enabled="true"/>
-  <check class="org.scalastyle.scalariform.UppercaseLChecker"                level="error"   enabled="true"/>
-  <check class="org.scalastyle.scalariform.VarFieldChecker"                  level="warning" enabled="true"/>
-  <check class="org.scalastyle.file.RegexChecker"                            level="warning" enabled="true">
-    <parameters>
-      <parameter name="regex">println</parameter>
-    </parameters>
-  </check>
-</scalastyle>
