--- core/src/sbt-test/scalastyle/compile/test
@@ -1,1 +1,0 @@
--> checkStyle
