--- core/src/sbt-test/scalastyle/ok/test
@@ -1,2 +1,0 @@
-> checkStyle
-> test:checkStyle
