--- core/src/sbt-test/scalastyle/test/project/plugins.sbt
@@ -1,1 +1,0 @@
-addSbtPlugin("com.nrinaudo" % "kantan.sbt" % sys.props("plugin.version"))
