--- core/src/sbt-test/scalastyle/test/test
@@ -1,1 +1,0 @@
--> test:checkStyle
