+++ kantan/src/main/scala/kantan/sbt/kantan/KantanKantanPlugin.scala
@@ -20,7 +20,7 @@ import com.typesafe.sbt.SbtGit.git
 import com.typesafe.sbt.pgp.PgpKeys.publishSigned
 import kantan.sbt.KantanRelease
 import kantan.sbt.KantanScalafmtPlugin, KantanScalafmtPlugin.autoImport._
-import kantan.sbt.KantanScalastylePlugin, KantanScalastylePlugin.autoImport._
+import kantan.sbt.scalastyle.KantanScalastylePlugin, KantanScalastylePlugin.autoImport._
 import sbt._, Keys._
 import sbtrelease.ReleasePlugin, ReleasePlugin.autoImport._, ReleaseTransformations._
