+++ scalastyle/src/sbt-test/scalastyle/main/src/main/scala/BadlyFormed.scala
