+++ scalastyle/src/sbt-test/scalastyle/ok/src/main/scala/WellFormed.scala
