+++ scalastyle/src/sbt-test/scalastyle/ok/src/test/scala/NotBadlyFormed.scala
