+++ scalastyle/src/sbt-test/scalastyle/test/project/plugins.sbt
@@ -1,0 +1,2 @@
+addSbtPlugin(
+  "com.nrinaudo" % "kantan.sbt-scalastyle" % sys.props("plugin.version"))
