+++ scalastyle/src/sbt-test/scalastyle/test/src/test/scala/BadlyFormed.scala
